requirements.txt
altair==5.4.1
matplotlib==3.8.1
numpy==2.1.2
pandas==2.2.3
Pillow==10.1.0
Pillow==10.4.0
seaborn==0.13.2
streamlit==1.39.0
